requirements.txt
# Python dependencies for RAG application
# ========================================
# Install with: pip install -r requirements.txt

# Document processing
pypdf>=4.0.0          # PDF parsing (newer, faster than PyPDF2)
python-docx>=1.1.0    # Word document parsing

# Embeddings & LLM
google-generativeai>=0.8.0  # Google Gemini API
google.genai

# Vector database
weaviate-client>=4.9.0  # Weaviate Python client (v4)

# Utilities
python-dotenv>=1.0.0   # Load .env file for API keys

# Web Framework
flask>=3.0.0           # Web server
flask-cors>=4.0.0      # CORS support for API
